milestone upto 2

+++ FarmVerse/farmVerseApplication/src/main/java/edu/infosys/farmVerseApplication/dao/CropDaoImpl.java
@@ -44,7 +44,7 @@ public class CropDaoImpl implements CropDao {
 		}
 
 		@Override
-		public Long getMaxCropId() {
+		public Integer getMaxCropId() {
 			
 			return repository.getMaxCropId();
 		}
